FEA/MFEA refactoring: get rid of the raw sockets I/O inside the MFEA, and use the FEA instead.
Modify accordingly the MLD/IGMP and PIM-SM modules to use the FEA
instead for raw I/O.

Note that after the modification only PIM-SM needs to register
with the MFEA (per interface) for the purpose of multicast routing.
Also, unlike before there could be only one module/target that
can register (as an owner) per interface with the MFEA.
This is also a step toward implementing a module with the
interoperability rules for multicast routing protocols
(as per RFC 2715).

Note that the refactored code doesn't contain anymore the hack
for sharing the IGMP raw socket and the (MRT_INIT) multicast routing
socket. That hack was needed because Linux seems to use
the MRT_INIT multicast routing socket (only) to pass IGMP messages
to userland (if the IGMP messages don't contain the Router Alert option
and the host hasn't joined the multicast group for the IGMP destination
address).
This issue needs further investigation and eventually will be fixed
independently.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10039 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/fea/rawsock4.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/fea/rawsock4.cc,v 1.18 2007/02/16 22:45:49 pavlin Exp $"
+#ident "$XORP: xorp/fea/rawsock4.cc,v 1.19 2007/05/08 19:23:14 pavlin Exp $"
 
 #include "fea_module.h"
 
@@ -26,7 +26,7 @@
 /* ------------------------------------------------------------------------- */
 /* RawSocket4 methods */
 
-RawSocket4::RawSocket4(EventLoop& eventloop, uint32_t ip_protocol,
+RawSocket4::RawSocket4(EventLoop& eventloop, uint8_t ip_protocol,
 		       const IfTree& iftree)
     : RawSocket(eventloop, IPv4::af(), ip_protocol, iftree)
 {
